requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
openai==1.75.0
PyPDF2==3.0.1
python-dotenv==1.0.1
python-multipart==0.0.20
